requirements.txt
Flask==0.9
Flask-Assets==0.8
Jinja2==2.6
Werkzeug==0.7.1
psycopg2==2.4.5
django-htmlmin==0.5.2
webassets==0.8
jsmin==2.0.2
cssmin==0.1.4
